changes from review
moved the caching of target position for a salvo wave to be at first delivery and not at scheduling time, as this is how the single salvo used to work. Required bunch of changes to facilitate.
Undid the explosive creation position change as that was only where they spawn and not where the land. Added comments to this area of code to clarify.
Added minor clarifications to the readme.
Let vscode auto standardise formatting of explosive-delivery
Resolve old (unrelated) EmmyLua warnings in other action type script files.