Review and clarify Teleport & Call For Help
No functional changes.
Tiny caching optimisation.
Note removed as path request needs to use the entity prototypes collision box size and not the specific entity instances bounding box in the world map.
Lots of edge cases identified from code review of Teleport and Call For Help due to the async nature of waiting for path finder requests to return and the possible changes to game state during this time.