testing and release preperation
TODO around aggressive driver to be given a vehicle changed to be a separate feature as that is easier and it is technically a generic feature.
Fix looping code in direction-utils. It was copied from separate looping int code in math utils. Only affected numbers outside of the expected range (not present in this streamer mod).