Inventory Drop UPS Logic - WIP
move from searching through the cached item counts for each item to drop, to getting the item numbers to be dropped and then working through the cached item counts to find the next numbered item to be dropped. Enabled by the placement logic change to being random angle and distance, as items no longer need to be dropped in a random order, just random items is good enough.